refactor(bot): move common SeriesParser methods to BaseSeriesParser
